pyproject.toml
[project]
name = "llamaindex-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index",
    "llama-index-llms-zhipuai",
    "llama-index-embeddings-zhipuai",
    "llama-index-vector-stores-chroma>=0.2.0",
    "python-dotenv",
    "chromadb>=0.4.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]
